+++ web/satellite/src/components/header/AccountDropdown.vue
@@ -34,11 +34,11 @@
         APP_STATE_ACTIONS,
         PROJETS_ACTIONS,
         PM_ACTIONS,
-        USER_ACTIONS,
         API_KEYS_ACTIONS,
         NOTIFICATION_ACTIONS,
         BUCKET_USAGE_ACTIONS,
     } from '@/utils/constants/actionNames';
+    import { USER_ACTIONS } from '@/store/modules/users';
 
     @Component
     export default class ProjectSelectionDropdown extends Vue {
