+++ web/satellite/tests/unit/store/credits.spec.ts
@@ -5,9 +5,8 @@ import Vuex from 'vuex';
 import { createLocalVue } from '@vue/test-utils';
 import { CreditsApiGql } from '@/api/credits';
 import { makeCreditsModule } from '@/store/modules/credits';
-import { CREDIT_USAGE_MUTATIONS } from '@/store/mutationConstants';
+import { CREDIT_USAGE_MUTATIONS, CREDIT_USAGE_ACTIONS } from '@/store/modules/credits';
 import { CreditUsage } from '@/types/credits';
-import { CREDIT_USAGE_ACTIONS } from '@/utils/constants/actionNames';
 
 const Vue = createLocalVue();
 const api = new CreditsApiGql();
