+++ web/satellite/tests/unit/store/users.spec.ts
@@ -5,9 +5,8 @@ import Vuex from 'vuex';
 import { createLocalVue } from '@vue/test-utils';
 import { UsersApiGql } from '@/api/users';
 import { makeUsersModule } from '@/store/modules/users';
-import { USER_MUTATIONS } from '@/store/mutationConstants';
 import { UpdatedUser, User } from '@/types/users';
-import { USER_ACTIONS } from '@/utils/constants/actionNames';
+import { USER_MUTATIONS, USER_ACTIONS } from '@/store/modules/users';
 
 const Vue = createLocalVue();
 const usersApi = new UsersApiGql();
@@ -81,6 +80,16 @@ describe('actions', () => {
             expect(store.state.fullName).toBe(oldUser.fullName);
             expect(store.state.shortName).toBe(oldUser.shortName);
         }
+    });
+
+    it('clears state', async () => {
+        await store.dispatch(CLEAR);
+
+        expect(store.state.fullName).toBe('');
+        expect(store.state.shortName).toBe('');
+        expect(store.state.email).toBe('');
+        expect(store.state.partnerId).toBe('');
+        expect(store.state.id).toBe('');
     });
 
     it('success get user', async () => {
